inline distributed helpers into rfdetr loss
